+++ Gruntfile.js
@@ -86,7 +86,6 @@ module.exports = function(grunt) {
       chrome: {
         options: {
           args: {
-            browser: 'chrome'
           }
         }
       }