+++ source/autoCompBase.js
@@ -591,6 +591,10 @@ if (typeof Def === 'undefined')
      *     item in the list.</li>
      *    <li>maxSelect - (default 1) The maximum number of items that can be
      *     selected.  Use '*' for unlimited.</li>
+     *    <li>tokens - (default none) For autocompleting based on part of the
+     *     field's value, this should be an array of the characters that are
+     *     considered "word" boundaries (e.g. a space, but could be something
+     *     else).  When this option is used, maxSelect is ignored.
      *    <li>scrolledContainer - the element that should be scrolled to bring
      *     the list into view if it would otherwise extend below the edge of the
      *     window. The default is document.documentElement (i.e. the whole
@@ -622,9 +626,9 @@ if (typeof Def === 'undefined')
       if (this.twoColumnFlow_ === undefined)
         this.twoColumnFlow_ = true;
 
-      if (options.maxSelect === undefined)
+      if (options.tokens || options.maxSelect === undefined)
         options.maxSelect = 1;
-      if (options.maxSelect === '*')
+      else if (options.maxSelect === '*')
         options.maxSelect = Infinity;
       this.multiSelect_ = options.maxSelect !== 1;
       if (options.scrolledContainer !== undefined) // allow null
@@ -664,8 +668,17 @@ if (typeof Def === 'undefined')
       this.element.setAttribute('autocomplete','off');
       // --- end of section copied from controls.js baseInitialize ---
       jQuery(this.update).hide();
-      jQuery(this.element).blur(jQuery.proxy(this.onBlur, this));
-      jQuery(this.element).keydown(jQuery.proxy(this.onKeyPress, this));
+      var jqElem = jQuery(this.element);
+      jqElem.blur(jQuery.proxy(this.onBlur, this));
+      jqElem.keydown(jQuery.proxy(this.onKeyPress, this));
+
+      // On clicks, reset the token bounds relative to the point of the click
+      if (this.options.tokens) {
+        jqElem.click(function () {
+          this.tokenBounds = null;
+          this.getTokenBounds(this.element.selectionStart);
+        }.bind(this));
+      }
 
       // If this is a multiselect list, put the field into a span.
       if (options.maxSelect > 1) {
@@ -715,6 +728,7 @@ if (typeof Def === 'undefined')
       this.index = -1;
 
       this.initDOMCache();
+      this.oldElementValue = this.domCache.get('elemVal');
     },
 
 
@@ -731,7 +745,8 @@ if (typeof Def === 'undefined')
       var fieldVal;
       if (runChangeEventObservers)
         fieldVal = this.domCache.get('elemVal');
-      this.domCache.set('elemVal', this.element.value = val);
+      this.domCache.set('elemVal', this.element.value = this.oldElementValue = val);
+      this.tokenBounds = null;
       if (runChangeEventObservers && fieldVal !== val) {
         Def.Event.simulate(this.element, 'change');
       }
@@ -1014,11 +1029,7 @@ if (typeof Def === 'undefined')
 
 
       this.scrollToShow(highlightedLITag, this.update.parentNode);
-
-      // Also put the value into the field, but don't run the change event yet,
-      // because the user has not really selected it.
-      this.setFieldVal(this.listItemValue(highlightedLITag), false);
-      this.element.select();
+      this.updateElementAfterMarking(highlightedLITag);
     },
 
 
@@ -1047,11 +1058,27 @@ if (typeof Def === 'undefined')
 
 
       this.scrollToShow(highlightedLITag, this.update.parentNode);
+      this.updateElementAfterMarking(highlightedLITag);
+    },
 
+
+    /**
+     *  Updates the field after an element has been highlighted in the list
+     *  (e.g. via arrow keys).
+     * @param listElement the DOM element that has been highlighted
+     */
+    updateElementAfterMarking: function(listElement) {
       // Also put the value into the field, but don't run the change event yet,
       // because the user has not really selected it.
-      this.setFieldVal(this.listItemValue(highlightedLITag), false);
-      this.element.select();
+      var oldTokenBounds = this.tokenBounds;
+      this.updateElement(listElement);  // clears this.tokenBounds
+      if (this.options.tokens) {
+        // Recompute token bounds, because we've inserted a list value
+        this.getTokenBounds(oldTokenBounds && oldTokenBounds[0]);
+        this.element.setSelectionRange(this.tokenBounds[0], this.tokenBounds[1]);
+      }
+      else
+        this.element.select();
     },
 
 
@@ -1171,7 +1198,6 @@ if (typeof Def === 'undefined')
      * @param event the event object from the keypress event
      */
     onKeyPress: function(event) {
-      this.domCache.invalidate('elemVal');
       // Do nothing if the autocompleter widget is not enabled_.
       if (this.enabled_) {
         // Note:  Normal (i.e. not search or navigation) key strokes are handled
@@ -1248,6 +1274,10 @@ if (typeof Def === 'undefined')
                           break;
                         case keys.LEFT:
                         case keys.RIGHT:
+                          if (this.options.tokens) {
+                            this.tokenBounds = null; // selection point may have moved
+                            this.getTokenBounds(); // selection point may have moved
+                          }
                           if (!event.ctrlKey && this.index>=0 &&
                               jQuery(this.update).hasClass('multi_col')) {
                             this.moveToOtherColumn(event);
@@ -1817,23 +1847,62 @@ if (typeof Def === 'undefined')
 
 
     /**
-     *  Since we aren't using the tokenizing stuff (which allows more than
-     *  one selection from the list) we are overriding getToken to just
-     *  return the element's full value.
+     *  Returns the part of the field value (maybe the full field value) that
+     *  should be used as the basis for autocompletion.
      */
     getToken: function() {
-      return this.domCache.get('elemVal');
+      var rtn = this.domCache.get('elemVal');
+      if (this.options.tokens) {
+        var bounds = this.getTokenBounds();
+        rtn = rtn.substring(bounds[0], bounds[1]);
+      }
+      return rtn;
     },
 
 
     /**
-     *  Since we aren't using the tokenizing stuff (which allows more than
-     *  one selection from the list) we are overriding getTokenBounds to just
-     *  return the bounds of the element's full value.
+     *  Returns the indices of the most recently changed part of the element's
+     *  value whose boundaries are the closest token characters.  Use when
+     *  autocompleting based on just part of the field's value.  Note that the
+     *  value is cached.  If you want an updated value, clear this.tokenBounds.
+     * @param pos (optional) a position in the string around which to extract
+       * the token.  Used when the changed part of the string is not known, or
+       * when there is no changed part but the user has clicked on a token.
      */
-    getTokenBounds: function() {
-      return [0, this.domCache.get('elemVal').length];
-    },
+    getTokenBounds: (function() {
+      /*
+         This function was used in Scriptaculous, but we are not basing the
+         concept of current tokens on what has changed, but on where the
+         cursor is in the field.  Retaining for referrence in case we need it.
+      function getFirstDifferencePos(newS, oldS) {
+        var boundary = Math.min(newS.length, oldS.length);
+        for (var index = 0; index < boundary; ++index)
+          if (newS[index] != oldS[index])
+            return index;
+        return boundary;
+      };
+      */
+
+      return function(pos) {
+        if (null != this.tokenBounds) return this.tokenBounds;
+        var value = this.domCache.get('elemVal');
+        if (value.trim() === '') return [-1, 0];
+        // diff = position around which a token will be found.
+        var diff =  pos !== undefined ? pos : this.element.selectionStart;
+        // var diff = pos !== undefined ? pos :
+        //  getFirstDifferencePos(value, this.oldElementValue);
+        var offset = (diff == this.oldElementValue.length ? 1 : 0);
+        var prevTokenPos = -1, nextTokenPos = value.length;
+        var tp;
+        for (var index = 0, l = this.options.tokens.length; index < l; ++index) {
+          tp = value.lastIndexOf(this.options.tokens[index], diff + offset - 1);
+          if (tp > prevTokenPos) prevTokenPos = tp;
+          tp = value.indexOf(this.options.tokens[index], diff + offset);
+          if (-1 != tp && tp < nextTokenPos) nextTokenPos = tp;
+        }
+        return (this.tokenBounds = [prevTokenPos + 1, nextTokenPos]);
+      }
+    })(),
 
 
     /**
@@ -2229,6 +2298,8 @@ if (typeof Def === 'undefined')
         if (this.multiSelect_)
           this.showList();
       }
+
+      this.tokenBounds = null; // selection point may have moved
     },
 
 
@@ -2435,13 +2506,23 @@ if (typeof Def === 'undefined')
 
     /**
      *  Updates the field with the selected list item value.
-     * @param selectedElement the DOM LI element the user selected.
+     * @param selectedElement the DOM list element (LI or TR) the user selected.
      */
     updateElement: function(selectedElement) {
-      // The Scriptaculous autocompleters allow you to autocomplete more than
-      // once in a field and select more than one value from the list.  We're
-      // not doing that, so we don't do the getTokenBounds() stuff.
-      this.setFieldVal(this.listItemValue(selectedElement), false);
+      var selectedVal = this.listItemValue(selectedElement);
+      var newFieldVal = selectedVal;
+      if (this.options.tokens) { // We're autocompleting on paritial field values
+        var bounds = this.getTokenBounds();
+        if (bounds[0] != -1) {
+          var currentVal = this.domCache.get('elemVal');
+          var newValue = currentVal.substr(0, bounds[0]);
+          var whitespace = currentVal.substr(bounds[0]).match(/^\s+/);
+          if (whitespace)
+            newValue += whitespace[0];
+          newFieldVal = newValue + selectedVal + currentVal.substr(bounds[1]);
+        }
+      }
+      this.setFieldVal(newFieldVal, false);
       // The "false" argument above means do not run change observers.  After
       // this gets called, propagateFieldChanges is called, and that takes care
       // of running change event handlers.
@@ -2449,7 +2530,6 @@ if (typeof Def === 'undefined')
       if (this.options.afterUpdateElement)
         this.options.afterUpdateElement(this.element, selectedElement);
     },
-
 
 
     /**
@@ -2527,8 +2607,8 @@ if (typeof Def === 'undefined')
     },
 
 
-    // Copied as-is from controls.js  (remove this comment if you modify it).
     onObserverEvent: function() {
+      this.domCache.invalidate('elemVal'); // presumably the field value changed
       this.changed = false;
       this.tokenBounds = null;
       if(this.getToken().length>=this.options.minChars) {